Separate some preview functions into utils
